Move support system specs into directory
As they're growing quite big it makes sense to keep these in separate
files.
