+++ spec/system/support_interface/countries_spec.rb
@@ -1,21 +1,9 @@
 require "rails_helper"
 
-RSpec.describe "Support", type: :system do
-  before { given_countries_exist }
+RSpec.describe "Countries support", type: :system do
+  it "allows modifying countries" do
+    given_countries_exist
 
-  it "features" do
-    when_i_am_authorized_as_a_support_user
-    when_i_visit_the_feature_flags_page
-    then_i_see_the_feature_flags
-
-    when_i_activate_the_service_open_flag
-    then_the_service_open_flag_is_on
-
-    when_i_deactivate_the_service_open_flag
-    then_the_service_open_flag_is_off
-  end
-
-  it "countries" do
     when_i_am_authorized_as_a_support_user
     when_i_visit_the_countries_page
     then_i_see_the_countries
@@ -50,39 +38,11 @@ RSpec.describe "Support", type: :system do
     create(:region, :national, country: create(:country, code: "CY"))
   end
 
-  def then_i_see_the_feature_flags
-    expect(page).to have_current_path("/support/features")
-    expect(page).to have_title("Features")
-    expect(page).to have_content("Features")
-  end
-
-  def then_the_service_open_flag_is_off
-    expect(page).to have_content("Feature “Service open” deactivated")
-    expect(page).to have_content("Service open\n- Inactive")
-  end
-
-  def then_the_service_open_flag_is_on
-    expect(page).to have_content("Feature “Service open” activated")
-    expect(page).to have_content("Service open\n- Active")
-  end
-
-  def when_i_activate_the_service_open_flag
-    click_on "Activate Service open"
-  end
-
   def when_i_am_authorized_as_a_support_user
     page.driver.basic_authorize(
       ENV.fetch("SUPPORT_USERNAME", "test"),
       ENV.fetch("SUPPORT_PASSWORD", "test")
     )
-  end
-
-  def when_i_deactivate_the_service_open_flag
-    click_on "Deactivate Service open"
-  end
-
-  def when_i_visit_the_feature_flags_page
-    visit support_interface_features_path
   end
 
   def when_i_visit_the_countries_page
